icp_agents/requirements.txt
requests
python-dotenv
pyyaml
gradio-client